fix: don't delete the internal map when closing a subgraph
A subgraph shares its internal map (from IDs to names) with its parent graph.
Thus closing its internal map frees not only its entries in the map, but the
parent's entries as well. Prior to MR !1489 (the fix to issue #1767) [0], this
was OK because looking up the name associated with an ID (`agnameof`), falls
back on constructing a local name [1] based on the ID. I.e. looking up ID 3 that
was not in the internal map would return the name "%3" which conveniently
matched the name that was previously in the internal map.

Following the !1489 changes, non-local names are in the internal map and are
managed there. A side effect of this is that any ID passed to `agnameof` is
expected to match an existing entry in the internal map. Most relevant to the
issue resolved in this commit, there is no longer any logic to assume certain
IDs can be cast to pointers to recover the string associated with them; IDs are
expected to be anonymous tokens, only usable for looking up the internal map.

The fdp layout algorithm triggered a code path that relied on the former
behavior. It allowed the internal map to be deleted, under the assumption it
could recover all the contained information through `agnameof` lookups on an
empty map. We repair this by preserving the internal map across artificial
subgraph construction and deletion. Fixes #1876, #1877.

  [0]: Specifically, the relevant commits that altered this behavior were:
    * 4f283dd1c02a6a4999b53ad2fcbf2264a7074a8b
    * 85b09cf13179b0e5ab8bddb4857e3d2af0a39a31
    * 9409324489a69557229d3d6f505857b9af85a913
    * 2a9449a99b2a2146fce01fa1d9713e999ad3dd4e
    * 14be5169ef49faad0f30cd9d36cdd438e1739f77
  [1]: Local names are the ones Graphviz constructs internally for nodes that do
    not have a user-provided name. They begin with '%'. See LOCALNAMEPREFIX and
    its uses.